feat: Implement TransformerV2 for enhanced status data processing and replace deprecated Transformer
